==> scripts/4-modeling.Rmd <==
---
title: "Modeling"
author: "Heqing Sun"
date: "6/18/2020"
output: html_document
---

## Environment Setup
```{r}
getwd()

# If this errors out, need to install packages
source('0-setup_environment.R')
```

## Access the RDS file from 3-data_wrangling
```{r}
train_data <- readRDS('../data/clean/train_data.rds')
test_data <- readRDS('../data/clean/test_data.rds')

X_train <- readRDS('../data/clean/X_train_imputed.rds')
X_test <- readRDS('../data/clean/X_test_imputed.rds')
y_train <- readRDS( '../data/clean/y_train.rds')
y_test <- readRDS('../data/clean/y_test.rds')

# Convert target variable to factor
train_data$sale <- as.factor(train_data$sale)
test_data$sale <- as.factor(test_data$sale)
```

## Correlation Matrix
```{r}
m.res <- cor(X_train[sapply(X_train, is.numeric)], use='pairwise')

corrplot(m.res, type = "upper", order = "hclust", 
         tl.col = "black", tl.srt = 45)
```

## Modeling the original imbalanced data
```{r}
set.seed(123)
model_lr <- caret::train(sale ~ .,
                         data = train_data,
                         method = "glm",
                         preProcess = c("scale", "center"),
                         trControl = trainControl(method = "repeatedcv", 
                                                  number = 10, 
                                                  repeats = 10, 
                                                  verboseIter = FALSE))

final <- data.frame(actual = test_data$sale,
                    predict(model_lr, newdata = test_data, type = "prob"))
final$predict <- ifelse(final$X0 > 0.5, 0, 1)

cm_original <- confusionMatrix(factor(final$predict), factor(final$actual), positive = "1")
cm_original
#           Reference
# Prediction    0    1
#          0 4955  856
#          1   34   61
```

# Under-Sampling
```{r}
set.seed(123)
ctrl <- trainControl(method = "repeatedcv", 
                     number = 10, 
                     repeats = 10, 
                     verboseIter = FALSE,
                     sampling = "down")

model_lr_under <- caret::train(sale ~ .,
                         data = train_data,
                         method = "glm",
                         preProcess = c("scale", "center"),
                         trControl = ctrl)

final_under <- data.frame(actual = test_data$sale,
                    predict(model_lr_under, newdata = test_data, type = "prob"))
final_under$predict <- ifelse(final_under$X0 > 0.5, 0, 1)

cm_under <- confusionMatrix(factor(final_under$predict), factor(final_under$actual), positive = "1")
cm_under
#           Reference
# Prediction    0    1
#          0 3317  404
#          1 1672  513
```

## Over-Sampling
```{r}
set.seed(123)
ctrl <- trainControl(method = "repeatedcv", 
                     number = 10, 
                     repeats = 10, 
                     verboseIter = FALSE,
                     sampling = "up")

model_lr_over <- caret::train(sale ~ .,
                         data = train_data,
                         method = "glm",
                         preProcess = c("scale", "center"),
                         trControl = ctrl)

final_over <- data.frame(actual = test_data$sale,
                          predict(model_lr_over, newdata = test_data, type = "prob"))
final_over$predict <- ifelse(final_over$X0 > 0.5, 0, 1)

cm_over <- confusionMatrix(factor(final_over$predict), factor(final_over$actual), positive = "1")
cm_over
#           Reference
# Prediction    0    1
#          0 3317  402
#          1 1672  515
```

## ROSE
```{r}
set.seed(123)
ctrl <- trainControl(method = "repeatedcv", 
                     number = 10, 
                     repeats = 10, 
                     verboseIter = FALSE,
                     sampling = "rose")

model_lr_rose <- caret::train(sale ~ .,
                              data = train_data,
                              method = "glm",
                              preProcess = c("scale", "center"),
                              trControl = ctrl)

final_rose <- data.frame(actual = test_data$sale,
                          predict(model_lr_rose, newdata = test_data, type = "prob"))
final_rose$predict <- ifelse(final_rose$X0 > 0.5, 0, 1)

cm_rose <- confusionMatrix(factor(final_rose$predict), factor(final_rose$actual), positive = "1")
cm_rose
#           Reference
# Prediction    0    1
#          0 3313  403
#          1 1676  514
```

## ROSE Model Results
```{r}
model_lr_rose
model_lr_rose$finalModel
summary(model_lr_rose)
```


## SMOTE (Synthetic Minority Oversampling Technique)
```{r}
set.seed(123)
ctrl <- trainControl(method = "repeatedcv", 
                     number = 10, 
                     repeats = 10, 
                     verboseIter = FALSE,
                     sampling = "smote")

model_lr_smote <- caret::train(sale ~ .,
                              data = train_data,
                              method = "glm",
                              preProcess = c("scale", "center"),
                              trControl = ctrl)

final_smote <- data.frame(actual = test_data$sale,
                          predict(model_lr_smote, newdata = test_data, type = "prob"))
final_smote$predict <- ifelse(final_smote$X0 > 0.5, 0, 1)

cm_smote <- confusionMatrix(factor(final_smote$predict), factor(final_smote$actual), positive = "1")
cm_smote
#           Reference
# Prediction    0    1
#          0 3627  473
#          1 1362  444
```

## Compare the predictions of all these models - boxplot
```{r}
models <- list(original = model_lr,
                       under = model_lr_under,
                       over = model_lr_over,
                       smote = model_lr_smote,
                       rose = model_lr_rose)

resampling <- resamples(models)
bwplot(resampling)

## SMOTE have the comparable accuuracy to original, but way higher specificity and sensitivity
```

## Compare the predictions of all these models
```{r}
comparison <- data.frame(model = names(models),
                         Precision = rep(NA, length(models)),
                         Recall = rep(NA, length(models)),
                         F1 = rep(NA, length(models)))

comparison[comparison$model == "original", ] <- filter(comparison, model == "original") %>%
    mutate(Precision = cm_original$byClass["Precision"],
           Recall = cm_original$byClass["Recall"],
           F1 = cm_original$byClass["F1"])

comparison[comparison$model == "under", ] <- filter(comparison, model == "under") %>%
    mutate(Precision = cm_under$byClass["Precision"],
           Recall = cm_under$byClass["Recall"],
           F1 = cm_under$byClass["F1"])

comparison[comparison$model == "over", ] <- filter(comparison, model == "over") %>%
    mutate(Precision = cm_over$byClass["Precision"],
           Recall = cm_over$byClass["Recall"],
           F1 = cm_over$byClass["F1"])

comparison[comparison$model == "smote", ] <- filter(comparison, model == "smote") %>%
    mutate(Precision = cm_smote$byClass["Precision"],
           Recall = cm_smote$byClass["Recall"],
           F1 = cm_smote$byClass["F1"])

comparison[comparison$model == "rose", ] <- filter(comparison, model == "rose") %>%
    mutate(Precision = cm_rose$byClass["Precision"],
           Recall = cm_rose$byClass["Recall"],
           F1 = cm_rose$byClass["F1"])

## FOR LOOP not working here
# for (name in names(models)) {
#   model <- get(paste0("cm_", name))
# 
#   comparison[comparison$model == name, ] <- filter(comparison, model == name) %>%
#     mutate(Sensitivity = model$byClass["Sensitivity"],
#            Specificity = model$byClass["Specificity"],
#            Precision = model$byClass["Precision"],
#            Recall = model$byClass["Recall"],
#            F1 = model$byClass["F1"])
# }

comparison %>%
  gather(x, y, Precision:F1) %>%
  ggplot(aes(x = x, y = y, color = model)) +
    geom_jitter(width = 0.2, alpha = 0.5, size = 3)
```
