==> scripts/5-part2&3_calculations.Rmd <==
---
title: "Part 2&3 Calculations"
author: "Heqing Sun"
date: "6/18/2020"
output: html_document
---

## Environment Setup
```{r}
getwd()

# If this errors out, need to install packages
source('0-setup_environment.R')
```

## Access the RDS file from previous steps
```{r}
train_data <- readRDS('../data/clean/train_data.rds')
test_data <- readRDS('../data/clean/test_data.rds')
data_sub_mi <- readRDS('../data/clean/data_sub_mi.rds')
```

## Data Preparation
```{r}
# Use the same partition method to split the data_sub_mi to get test data index
set.seed(123)
m.smp_size <- floor(0.75 * nrow(data_sub_mi))
m.train_ind <- sample(seq_len(nrow(data_sub_mi)), size = m.smp_size)
m.test <- data_sub_mi[-m.train_ind, ]

# Set Ref1 as the index
rownames(test_data) <- rownames(m.test)
```

## Use ROSE Logistic Regression
```{r}
# Convert target variable to factor
train_data$sale <- as.factor(train_data$sale)
test_data$sale <- as.factor(test_data$sale)

set.seed(123)
ctrl <- trainControl(method = "repeatedcv", 
                     number = 10, 
                     repeats = 10, 
                     verboseIter = FALSE,
                     sampling = "rose")

model_lr_rose <- caret::train(sale ~ .,
                              data = train_data,
                              method = "glm",
                              preProcess = c("scale", "center"),
                              trControl = ctrl)

final_rose <- data.frame(actual = test_data$sale,
                          predict(model_lr_rose, newdata = test_data, type = "prob"))
final_rose$predict <- ifelse(final_rose$X0 > 0.5, 0, 1)

cm_rose <- confusionMatrix(factor(final_rose$predict), factor(final_rose$actual), positive = "1")
cm_rose
```

## Plot Precision - Recall Curve
```{r}
# Get each class's probabilities
m.a <- final_rose %>% filter(actual==1)
m.pc <-  m.a$X1
m.b <- final_rose %>% filter(actual==0)
m.nc <-  m.b$X1

# PR Curve
pr <- pr.curve(scores.class0 = m.pc, scores.class1 = m.nc, curve = T)
plot(pr, auc.main=FALSE)

# Create a dateframe that contains thw detailed values
m.threshold <- data.frame(pr$curve)
colnames(m.threshold) <- c("recall", "precision", "threshold")
## Find that when the threshold >= 0.5832411, the precision will greater than 1/3
```

# Part 2 - Select Promising Records
```{r}
selected <- final_rose %>% filter(X1 >= 0.5832411) ## 447 obs
table(selected$actual)
#   0   1 
# 298 149 

selected
write.csv(selected,'../data/clean/selected_records_for_campaign.csv')

# Clean up environment
rm(list = ls(pattern = "^m."))
```

## Calculate the variables means for selected variables
```{r}
unselected <- final_rose %>% filter(X1 < 0.5832411) ## 5459 obs

# Get the index
m.unselected.index <- rownames(unselected)
m.selected.index <- rownames(selected)

# Subset dataframe into two parts
m.selected_data <- test_data[m.selected.index, ]
m.unselected_data <- test_data[m.unselected.index, ]

colMeans(m.selected_data)

numdata<-m.selected_data[sapply(m.selected_data, is.numeric)]  
lapply(numdata, mean, na.rm = T)  # Returns a list  

numdata<-m.unselected_data[sapply(m.unselected_data, is.numeric)]  
lapply(numdata, mean, na.rm = T)  # Returns a list  

# Create selected column
m.selected_data$selected <- 1
m.unselected_data$selected <- 0

# Join two dataframes vertically
m.test_withSelectedColumn <- rbind(m.selected_data, m.unselected_data)
m.test_withSelectedColumn_melted <- melt(m.test_withSelectedColumn, id.var = "selected")

```

## Calculate the differences between two groups
```{r}
names(m.test_withSelectedColumn)
ggplot(m.test_withSelectedColumn,aes(x=CUSTOMER_CANCEL_REASON_OTHER))+geom_histogram()+facet_grid(~selected)+theme_bw()
ggplot(m.test_withSelectedColumn,aes(x=OPTIONS_PRICE))+geom_histogram()+facet_grid(~selected)+theme_bw()
ggplot(m.test_withSelectedColumn,aes(x=BCC5421_AVG))+geom_histogram()+facet_grid(~selected)+theme_bw()
ggplot(m.test_withSelectedColumn,aes(x=BCC5320_AVG))+geom_histogram()+facet_grid(~selected)+theme_bw()
ggplot(m.test_withSelectedColumn,aes(x=BCC6200A_PCT))+geom_histogram()+facet_grid(~selected)+theme_bw()
ggplot(m.test_withSelectedColumn,aes(x=MTF6200A_PCT))+geom_histogram()+facet_grid(~selected)+theme_bw()
ggplot(m.test_withSelectedColumn,aes(x=CUSTOMER_CANCEL_REASON_U))+geom_histogram()+facet_grid(~selected)+theme_bw()
ggplot(m.test_withSelectedColumn,aes(x=BCC6160A_PCT))+geom_histogram()+facet_grid(~selected)+theme_bw()

ggplot(m.test_withSelectedColumn,aes(x=ILN7430_AVG))+geom_histogram()+facet_grid(~selected)+theme_bw()
ggplot(m.test_withSelectedColumn,aes(x=ALL6230A_PCT))+geom_histogram()+facet_grid(~selected)+theme_bw()
ggplot(m.test_withSelectedColumn,aes(x=BCC5421_AVG_missing))+geom_histogram()+facet_grid(~selected)+theme_bw()
ggplot(m.test_withSelectedColumn,aes(x=OPTIONS_PRICE_missing))+geom_histogram()+facet_grid(~selected)+theme_bw()
```

## Part 3 - Calculate the Sample Size
```{r}
# Calculate the sales rate in the test data
summary(test_data$sale)
## ~ 0.18

power.prop.test(p1=0.18, p2=0.18*1.05, power=0.8, alternative='two.sided', sig.level=0.1)
```

```{r}
# Calculate the sales rate in the test data
summary(test_data$sale)
## ~ 0.18
power.prop.test(p1=0.18, p2=0.18*1.05, power=0.8, alternative='two.sided', sig.level=0.05)
power.prop.test(p1=0.18, p2=0.18*1.05, power=0.8, alternative='one.sided', sig.level=0.1)
```
